+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # develop
   * Feature: Asterisk calls receiving media commands are implicitly answered
+  * Bugfix: Unrenderable output documents on Asterisk should return a sensible error
 
 # v0.9.1 - 2012-01-30
   * Bugfix: Closing an disconnected XMPP connection is a no-op